Complete rework of #include hierarchy
