+++ Constants.hpp
@@ -42,3 +42,31 @@
 // USB serial port speed according to external controller
 #define SERIAL_BAUDRATE_STELLARIUM_DIRECT   9600
 #define SERIAL_BAUDRATE_ASCOM               57600
+
+
+// Wifi operating modes (ESP32 only)
+#define WIFI_MODE_INFRASTRUCTURE                        0   // Infrastructure Only - OAT connects to an existing Router
+#define WIFI_MODE_AP_ONLY                               1   // AP Mode Only        - OAT acts as a local Router/Hotspot
+#define WIFI_MODE_ATTEMPT_INFRASTRUCTURE_FAIL_TO_AP     2   // Attempt Infrastructure, Fail over to AP Mode - Attempt infrastructure mode, with fail over to AP Mode.
+
+
+// Debugging output control
+// Each bit in the debug level specifies a kind of debug to enable. Combine these into DEBUG_LEVEL for the required information e.g.:
+//  #define DEBUG_LEVEL (DEBUG_STEPPERS|DEBUG_MOUNT)
+//  #define DEBUG_LEVEL (DEBUG_INFO|DEBUG_MOUNT|DEBUG_GENERAL)
+//  #define DEBUG_LEVEL (DEBUG_SERIAL|DEBUG_WIFI|DEBUG_INFO|DEBUG_MOUNT|DEBUG_GENERAL)
+//  #define DEBUG_LEVEL (DEBUG_ANY)
+//  #define DEBUG_LEVEL (DEBUG_INFO|DEBUG_MOUNT|DEBUG_GENERAL)
+#define DEBUG_NONE           0x0000     // No debug output (release build)
+#define DEBUG_INFO           0x0001     // General output, like startup variables and status
+#define DEBUG_SERIAL         0x0002     // Serial commands and replies
+#define DEBUG_WIFI           0x0004     // Wifi related output
+#define DEBUG_MOUNT          0x0008     // Mount processing output
+#define DEBUG_MOUNT_VERBOSE  0x0010     // Verbose mount processing (coordinates, etc)
+#define DEBUG_GENERAL        0x0020     // Other misc. output
+#define DEBUG_MEADE          0x0040     // Meade command handling output
+#define DEBUG_VERBOSE        0x0080     // High-rate mount activity, incl. stepper servicing
+#define DEBUG_STEPPERS       0x0100     // Stepper motor-related activity
+#define DEBUG_EEPROM         0x0200     // Storing and retrieval of non-volatile configuration data
+#define DEBUG_GYRO           0x0400     // Gyro activity (tilt/roll) calibration
+#define DEBUG_ANY            0xFFFF     // All debug output
