+++ platformio.ini
@@ -10,6 +10,7 @@
 
 [platformio]
 default_envs = mega2560
+include_dir = .
 
 [common]
 lib_deps =
