+++ src/Core.cpp
@@ -1,4 +1,8 @@
-#include "inc/Config.hpp"
+#include "inc/Globals.hpp"
+#include "../Configuration.hpp"
+#include "Utility.hpp"
+#include "Mount.hpp"
+
 #include "a_inits.hpp"
 #include "b_setup.hpp"
 #include "c65_startup.hpp"
