+++ src/LcdMenu.cpp
@@ -1,9 +1,10 @@
+#include "inc/Globals.hpp"
+#include "../Configuration.hpp"
 #include "Utility.hpp"
 #include "EPROMStore.hpp"
 #include "LcdMenu.hpp"
-#include "inc/Config.hpp"
 
-#if DISPLAY_TYPE > 0
+#if DISPLAY_TYPE != DISPLAY_TYPE_NONE
 
 // Class that drives the LCD screen with a menu
 // You add a string and an id item and this class handles the display and navigation
@@ -477,3 +478,4 @@ void LcdMenu::printChar(char ch) {}
 void LcdMenu::printAt(int col, int row, char ch) {}
 
 #endif
+
