+++ src/Longitude.hpp
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <Arduino.h>
 #include "DayTime.hpp"
 
 // -180..180 range, 0 is at the prime meridian (through Greenwich), negative going west, positive going east
@@ -14,7 +13,7 @@ public:
 
   virtual const char *formatString(char *targetBuffer, const char *format, long *pSeconds = nullptr) const;
 
-  static Longitude ParseFromMeade(String s);
+  static Longitude ParseFromMeade(String const& s);
 
 protected:
   virtual void checkHours() override;
