+++ src/Sidereal.cpp
@@ -1,6 +1,9 @@
+#include "../Configuration.hpp"
 #include "Sidereal.hpp"
 
 #if USE_GPS == 1
+
+#include <TinyGPS++.h>
 
 // Constants for sidereal calculation
 // Source: http://www.stargazing.net/kepler/altaz.html
