+++ src/a_inits.hpp
@@ -115,7 +115,6 @@ bool inStartup = false;        // Start with a guided startup
 #endif
 
 // Serial control variables
-bool inSerialControl = false; // When the serial port is in control
 bool okToUpdateMenu = true;   // Can be used to supress rendering the first line of the menu.
 bool quitSerialOnNextButtonRelease = false; // Used to detect SELECT button to quit Serial mode.
