--- src/inc/Config.hpp
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "../../Constants.hpp"
-#include "../../Configuration.hpp"
-#include "../../Configuration_adv.hpp"
-#include "../../Version.h"
