+++ src/inc/Globals.cpp
@@ -1,7 +1,4 @@
 #include "Globals.hpp"
-#include "Config.hpp"
 
-// TODO (AS): This is just a tmp fix to support menuCAL, we need to get rid of these globals
-
-int RAStepsPerDegree = RA_STEPS_PER_DEGREE * 10;       // menuCAL use higher resolution for editing
-int DECStepsPerDegree = DEC_STEPS_PER_DEGREE * 10;      // menuCAL uses higher resolution for editing
+// TODO: Should move this somewhere else...
+bool inSerialControl = false; // True when the serial port is in control
